+++ bcipy/helpers/demo/demo_eeg_model_related.py
@@ -1,5 +1,5 @@
 import numpy as np
-from bcipy.helpers.eeg_model_related import CopyPhraseWrapper
+from bcipy.helpers.signal_model_related import CopyPhraseWrapper
 from bcipy.signal.model.mach_learning.train_model import train_pca_rda_kde_model
 from bcipy.helpers.bci_task_related import alphabet
