+++ bcipy/helpers/signal_model_related.py
